Retrieve TestTable from AbstractPlannerTest.
